@@ -4,15 +4,12 @@
     "private": true,
     "dependencies": {
         "@7rulnik/react-loadable": "^5.3.4",
-        "ajv": "^6.4.0",
+        "ajv": "^6.5.0",
         "axios": "^0.18.0",
-        "fastify": "^1.2.1",
         "has": "^1.0.1",
         "lodash.flow": "^3.5.0",
         "lodash.isequal": "^4.5.0",
         "lodash.throttle": "^4.1.1",
-        "preact": "^8.2.7",
-        "preact-compat": "^3.18.0",
         "prop-types": "^15.6.1",
         "react": "^16.3.2",
         "react-dnd": "^2.6.0",
@@ -21,10 +18,8 @@
         "react-quill": "^1.2.7",
         "react-redux": "^5.0.7",
         "redux": "^4.0.0",
-        "redux-batched-actions": "^0.2.1",
-        "rmwc": "^1.5.6",
-        "serve-static": "^1.13.2",
-        "uws": "^9.14.0"
+        "redux-batched-actions": "^0.3.0",
+        "rmwc": "^1.6.3"
     },
     "devDependencies": {
         "@babel/core": "^7.0.0-beta.44",
@@ -32,49 +27,47 @@
         "@babel/plugin-proposal-decorators": "^7.0.0-beta.44",
         "@babel/plugin-proposal-object-rest-spread": "^7.0.0-beta.44",
         "@babel/plugin-syntax-dynamic-import": "^7.0.0-beta.44",
+        "@babel/plugin-transform-react-inline-elements": "^7.0.0-beta.47",
         "@babel/preset-env": "^7.0.0-beta.44",
         "@babel/preset-react": "^7.0.0-beta.44",
-        "autoprefixer": "^8.3.0",
+        "autoprefixer": "^8.5.0",
         "babel-eslint": "8.2.3",
         "babel-loader": "^8.0.0-beta.2",
         "babel-plugin-import": "^1.7.0",
         "clean-webpack-plugin": "^0.1.19",
-        "cross-env": "^5.1.4",
+        "cross-env": "^5.1.5",
         "css-loader": "^0.28.11",
         "cssnano": "^3.10.0",
         "eslint": "^4.19.1",
         "eslint-loader": "^2.0.0",
-        "eslint-plugin-import": "^2.11.0",
-        "eslint-plugin-react": "^7.7.0",
+        "eslint-plugin-import": "^2.12.0",
+        "eslint-plugin-react": "^7.8.2",
         "file-loader": "^1.1.11",
         "html-webpack-plugin": "^3.2.0",
         "husky": "^0.14.3",
         "mini-css-extract-plugin": "^0.4.0",
-        "node-sass": "^4.8.3",
-        "nodemon": "^1.17.3",
+        "node-sass": "^4.9.0",
         "optimize-js-plugin": "^0.0.4",
-        "postcss-loader": "^2.1.4",
+        "postcss-loader": "^2.1.5",
         "prettier": "^1.12.1",
-        "pretty-quick": "^1.4.1",
+        "pretty-quick": "^1.5.1",
         "sass-loader": "^7.0.1",
         "script-ext-html-webpack-plugin": "^2.0.1",
         "style-loader": "^0.21.0",
         "url-loader": "^1.0.1",
-        "webpack": "^4.6.0",
-        "webpack-bundle-analyzer": "^2.11.1",
-        "webpack-cli": "^2.0.14",
-        "webpack-dev-server": "^3.1.3",
+        "webpack": "^4.8.3",
+        "webpack-bundle-analyzer": "^2.13.0",
+        "webpack-cli": "^2.1.3",
+        "webpack-dev-server": "^3.1.4",
         "webpack-merge": "^4.1.2",
         "webpack-shell-plugin": "^0.5.0",
-        "write-file-webpack-plugin": "^4.2.0"
+        "write-file-webpack-plugin": "^4.3.2"
     },
     "scripts": {
         "start": "yarn run dev",
         "dev": "cross-env NODE_ENV=development webpack-dev-server",
         "build": "cross-env NODE_ENV=production webpack",
-        "server": "nodemon server.js",
-        "postinstall": "node scripts/symlink_source.js",
-        "precommit": "pretty-quick --staged"
+        "postinstall": "node scripts/symlink_source.js"
     },
     "prettier": {
         "tabWidth": 4
