@@ -1,22 +1,18 @@
-import React, { PureComponent } from 'react';
-import PropTypes from 'prop-types';
+import React, { PureComponent } from "react";
+import PropTypes from "prop-types";
 
-import styles from './styles.scss';
+import styles from "./styles.scss";
 
-export default class AppBar extends PureComponent {
+const AppBar = ({ title }) => (
+    <header className={styles.app_bar}>
+        <div className={styles.app_bar_content}>
+            <h2>{title}</h2>
+        </div>
+    </header>
+);
 
-    render() {
-        
-        return (
-            <header className={styles.app_bar}>
-                <div className={styles.app_bar_content}>
-                    <h2>{this.props.title}</h2>
-                </div>
-            </header>
-        )
-    }
+AppBar.propTypes = {
+    title: PropTypes.string.isRequired
+};
 
-    static propTypes = {
-        title: PropTypes.string.isRequired,
-    }
-}
+export default AppBar;
