@@ -1,67 +1,45 @@
-import React, { Component } from 'react';
-import ReactDOM  from 'react-dom';
-import PropTypes from 'prop-types';
+import React, { Component } from "react";
+import ReactDOM from "react-dom";
+import PropTypes from "prop-types";
 
-import styles from './styles.scss';
+import styles from "./styles.scss";
 
+const modal = document.getElementById("modal");
 
 export default class Modal extends Component {
-    constructor(props) {
-        super(props);
-
-        this.node = document.getElementById('modal');
-    }
-
     shouldComponentUpdate({ open }) {
-        return this.props.open !== open
-    }
-
-    componentWillUpdate({ open }) {
-        if(open) {
-            this.wrapper.classList.remove(`${styles.modal_wrapper_inactive}`);
-            this.wrapper.classList.add(`${styles.modal_wrapper_active}`);
-        } else {
-            this.wrapper.classList.remove(`${styles.modal_wrapper_active}`);
-            this.wrapper.classList.add(`${styles.modal_wrapper_inactive}`);
-        }
-
-        document.body.classList.toggle(styles.blur);
-    }
-
-    componentDidMount() {
-        if (open) {
-            this.wrapper.classList.remove(`${styles.modal_wrapper_inactive}`);
-            this.wrapper.classList.add(`${styles.modal_wrapper_active}`);
-        }
+        return this.props.open !== open;
     }
 
     render() {
-        const { 
-            open,
-            onClose,
-        } = this.props;
+        const { open, onClose } = this.props;
 
         return ReactDOM.createPortal(
-            <>
-                { <div 
-                    ref={(wrapper) => this.wrapper = wrapper} 
-                    className={`${styles.modal_wrapper}`} 
-                    onClick={onClose}
+            <div
+                className={`${styles.modal_wrapper} ${
+                    open
+                        ? styles.modal_wrapper_active
+                        : styles.modal_wrapper_inactive
+                }`}
+                onClick={onClose}
+            >
+                {open && (
+                    <div
+                        className={`${styles.modal} ${
+                            open ? styles.modal_open : ""
+                        } `}
                     >
-
-                    { <div className={`${styles.modal} ${open ? styles.modal_open : ''} `}>
-                        { this.props.children }
-                    </div>  }
-                </div>}
-            </>, this.node);
-    }
-
-    static defaultProps = {
+                        {this.props.children}
+                    </div>
+                )}
+            </div>,
+            modal
+        );
     }
 
     static propTypes = {
         open: PropTypes.bool.isRequired,
         onClose: PropTypes.func.isRequired,
-        children: PropTypes.node,
-    }
+        children: PropTypes.node
+    };
 }
